Generate consumer group in KafkaSourceUpdate test
Using the same consumer group for the different sources might produce
unreliable results because Kafka group ids are shared.

This is particularly evident in eventing-kafka-broker where the same
tests (`KafkaSourceUpdate`) is executed 5 times [1] in parallel [2]
(to catch flakes early), which means that there might be different
sources with the same consumer group pointing to 2 different sinks.

[1] https://github.com/knative-sandbox/eventing-kafka-broker/blob/e169e2cb923890a23e8d65651876292283b63908/test/e2e/kafka_source_test.go#L31
[2] https://github.com/knative-sandbox/eventing-kafka-broker/blob/e169e2cb923890a23e8d65651876292283b63908/test/pkg/testing/run.go#L52-L67

Signed-off-by: Pierangelo Di Pilato <pdipilat@redhat.com>